unify writability protections into a single handler
