+++ hbase-server/src/main/java/org/apache/hadoop/hbase/ipc/NettyRpcServerRequestDecoder.java
@@ -17,8 +17,6 @@
  */
 package org.apache.hadoop.hbase.ipc;
 
-import java.util.function.BooleanSupplier;
-import org.apache.hadoop.hbase.util.EnvironmentEdgeManager;
 import org.apache.hadoop.hbase.util.NettyFutureUtils;
 import org.apache.yetus.audience.InterfaceAudience;
 
@@ -36,41 +34,12 @@ class NettyRpcServerRequestDecoder extends SimpleChannelInboundHandler<ByteBuf> 
   private final MetricsHBaseServer metrics;
 
   private final NettyServerRpcConnection connection;
-  private final BooleanSupplier isWritabilityBackpressureEnabled;
-
-  private boolean writable;
-  private long unwritableStartTime;
 
   public NettyRpcServerRequestDecoder(MetricsHBaseServer metrics,
-    NettyServerRpcConnection connection, BooleanSupplier isWritabilityBackpressureEnabled) {
+    NettyServerRpcConnection connection) {
     super(false);
     this.metrics = metrics;
     this.connection = connection;
-    this.isWritabilityBackpressureEnabled = isWritabilityBackpressureEnabled;
-  }
-
-  @Override
-  public void channelWritabilityChanged(ChannelHandlerContext ctx) throws Exception {
-    // if backpressure is enabled, we want to manage the channel's setAutoRead based on
-    // whether the channel is writable. We also track metrics about how much time the channel
-    // spends unwritable.
-    if (isWritabilityBackpressureEnabled.getAsBoolean()) {
-      boolean oldWritableValue = this.writable;
-
-      this.writable = ctx.channel().isWritable();
-      ;
-      ctx.channel().config().setAutoRead(this.writable);
-
-      if (!oldWritableValue && this.writable) {
-        // changing from not writable to writable, update metrics
-        metrics.unwritableTime(EnvironmentEdgeManager.currentTime() - unwritableStartTime);
-        unwritableStartTime = 0;
-      } else if (oldWritableValue && !this.writable) {
-        // changing from writable to non-writable, set start time
-        unwritableStartTime = EnvironmentEdgeManager.currentTime();
-      }
-    }
-    ctx.fireChannelWritabilityChanged();
   }
 
   @Override
