+++ hbase-server/src/main/java/org/apache/hadoop/hbase/ipc/NettyRpcServerResponseEncoder.java
@@ -17,13 +17,9 @@
  */
 package org.apache.hadoop.hbase.ipc;
 
-import java.util.function.IntSupplier;
-import org.apache.hadoop.hbase.exceptions.ConnectionClosedException;
-import org.apache.hadoop.hbase.util.NettyUnsafeUtils;
 import org.apache.yetus.audience.InterfaceAudience;
 
 import org.apache.hbase.thirdparty.io.netty.buffer.Unpooled;
-import org.apache.hbase.thirdparty.io.netty.channel.Channel;
 import org.apache.hbase.thirdparty.io.netty.channel.ChannelHandlerContext;
 import org.apache.hbase.thirdparty.io.netty.channel.ChannelOutboundHandlerAdapter;
 import org.apache.hbase.thirdparty.io.netty.channel.ChannelPromise;
@@ -35,29 +31,17 @@ import org.apache.hbase.thirdparty.io.netty.channel.ChannelPromise;
 @InterfaceAudience.Private
 class NettyRpcServerResponseEncoder extends ChannelOutboundHandlerAdapter {
 
-  private static final ConnectionClosedException EXCEPTION =
-    new ConnectionClosedException("Channel outbound bytes exceeded fatal threshold");
   static final String NAME = "NettyRpcServerResponseEncoder";
 
   private final MetricsHBaseServer metrics;
-  private final IntSupplier pendingBytesFatalThreshold;
 
-  NettyRpcServerResponseEncoder(MetricsHBaseServer metrics,
-    IntSupplier pendingBytesFatalThreshold) {
+  NettyRpcServerResponseEncoder(MetricsHBaseServer metrics) {
     this.metrics = metrics;
-    this.pendingBytesFatalThreshold = pendingBytesFatalThreshold;
   }
 
   @Override
   public void write(ChannelHandlerContext ctx, Object msg, ChannelPromise promise)
     throws Exception {
-
-    // drop the message if fatal threshold is reached, as the connection will be closed
-    if (handleFatalThreshold(ctx)) {
-      promise.setFailure(EXCEPTION);
-      return;
-    }
-
     if (msg instanceof RpcResponse) {
       RpcResponse resp = (RpcResponse) msg;
       BufferChain buf = resp.getResponse();
@@ -70,29 +54,5 @@ class NettyRpcServerResponseEncoder extends ChannelOutboundHandlerAdapter {
     } else {
       ctx.write(msg, promise);
     }
-  }
-
-  private boolean handleFatalThreshold(ChannelHandlerContext ctx) {
-    int fatalThreshold = pendingBytesFatalThreshold.getAsInt();
-    if (fatalThreshold <= 0) {
-      return false;
-    }
-
-    Channel channel = ctx.channel();
-    long outboundBytes = NettyUnsafeUtils.getTotalPendingOutboundBytes(channel);
-    if (outboundBytes < fatalThreshold) {
-      return false;
-    }
-
-    if (channel.isOpen()) {
-      metrics.maxOutboundBytesExceeded();
-      RpcServer.LOG.warn(
-        "{}: Closing connection because outbound buffer size of {} exceeds fatal threshold of {}",
-        channel.remoteAddress(), outboundBytes, fatalThreshold);
-      NettyUnsafeUtils.closeImmediately(channel);
-    }
-
-    return true;
-
   }
 }
